+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 #Change Log
 
+##Version 1.1
+
+- Search for pending friend requests in the Followers and Following tabs of your profile page.
+- Temporary fix to autofocus text input field for messaging.
+- Updated updater to be more integrated into UI (user menu).
+- Updater now using GitHub API to check version.
+
 ##Version 1.0
 
 - Initial Tsu script that adds a bunch of tweaks to make Tsu more user friendly.